src/api/total_image/requirements.txt
numpy
opencv-python
tensorflow-gpu==1.8.0
Pillow
pathlib
keras
matplotlib
sklearn
easydict